+++ src/Waves.Core/GameContext/KruoGameContextBaseV2/KuroGameContextBaseV2.UpdateGame.cs
@@ -1,0 +1,554 @@
+﻿using Waves.Api.Models;
+using Waves.Core.Common;
+using Waves.Core.GameContext.KruoGameContextBaseV2.Common;
+using Waves.Core.Models;
+using Waves.Core.Models.Downloader;
+using Waves.Core.Models.Enums;
+
+namespace Waves.Core.GameContext;
+
+partial class KuroGameContextBaseV2
+{
+    #region 更新方法
+    public async Task<bool> UpdateGameResourceAsync()
+    {
+        var _launcher = await this.GetGameLauncherSourceAsync();
+        var currentVersion = await GameLocalConfig.GetConfigAsync(
+            GameLocalSettingName.LocalGameVersion
+        );
+        Task.Run(async () => await StartDownloadUpdateGameResourceAsync(_launcher, currentVersion));
+        return true;
+    }
+
+    /// <summary>
+    /// 预下载
+    /// </summary>
+    /// <returns></returns>
+    public async Task<bool> StartProdDownloadGameResourceAsync()
+    {
+        var _launcher = await this.GetGameLauncherSourceAsync();
+        var currentVersion = await GameLocalConfig.GetConfigAsync(
+            GameLocalSettingName.LocalGameVersion
+        );
+        Task.Run(async () =>
+            await StartDownloadUpdateGameResourceAsync(_launcher, currentVersion, true)
+        );
+        return true;
+    }
+
+    /// <summary>
+    /// 更新游戏
+    /// </summary>
+    /// <param name="_launcher"></param>
+    /// <param name="currentVersion"></param>
+    /// <param name="isProd"></param>
+    /// <returns></returns>
+    private async Task<bool> StartDownloadUpdateGameResourceAsync(
+        GameLauncherSource _launcher,
+        string currentVersion,
+        bool isProd = false
+    )
+    {
+        await _actionLock.WaitAsync();
+        try
+        {
+            #region 获取配置
+            if (_launcher == null || string.IsNullOrWhiteSpace(currentVersion))
+            {
+                GameEventPublisher.Publish(
+                    new GameContextOutputArgs()
+                    {
+                        Type = Models.Enums.GameContextActionType.TipMessage,
+                        TipMessage = "未找到更新配置文件，无法进行下载",
+                    }
+                );
+                return false;
+            }
+            var baseFolder = await this.GameLocalConfig.GetConfigAsync(
+                GameLocalSettingName.GameLauncherBassFolder
+            );
+            var previous = _launcher
+                .ResourceDefault.Config.PatchConfig.Where(x => x.Version == currentVersion)
+                .FirstOrDefault();
+            if (previous == null)
+            {
+                GameEventPublisher.Publish(
+                    new GameContextOutputArgs()
+                    {
+                        Type = Models.Enums.GameContextActionType.TipMessage,
+                        TipMessage = "未找到更新配置文件，无法进行下载",
+                    }
+                );
+                return false;
+            }
+            var cdnUrl =
+                _launcher
+                    .ResourceDefault.CdnList.Where(x => x.P != 0)
+                    .OrderBy(x => x.P)
+                    .FirstOrDefault() ?? null;
+            if (cdnUrl == null)
+            {
+                GameEventPublisher.Publish(
+                    new GameContextOutputArgs()
+                    {
+                        Type = Models.Enums.GameContextActionType.TipMessage,
+                        TipMessage = "未找到更新配置文件，无法进行下载",
+                    }
+                );
+                return false;
+            }
+            var _patch = await GetPatchGameResourceAsync(cdnUrl.Url + previous.IndexFile);
+            if (_patch == null)
+            {
+                GameEventPublisher.Publish(
+                    new GameContextOutputArgs()
+                    {
+                        Type = Models.Enums.GameContextActionType.TipMessage,
+                        TipMessage = "未找到更新配置文件，无法进行下载",
+                    }
+                );
+                return false;
+            }
+            #endregion
+
+            #region 初始化资源
+            this._downloadState = new DownloadState();
+
+            var downloadResource = new List<IndexResource>();
+            var patchResource = new List<IndexResource>();
+            var groupResource = new List<IndexResource>();
+            var zipResource = new List<IndexResource>();
+
+            foreach (var x in _patch.Resource)
+            {
+                if (x.Dest.Contains("krdiff"))
+                    patchResource.Add(x);
+                else if (x.Dest.Contains("krpdiff"))
+                    groupResource.Add(x);
+                else if (x.Dest.Contains("krzip"))
+                    zipResource.Add(x);
+                else
+                    downloadResource.Add(x);
+            }
+            string downloadBaseFolder = "";
+            if (isProd)
+            {
+                downloadBaseFolder = Path.Combine(baseFolder, "prodDownloads");
+            }
+            else
+            {
+                downloadBaseFolder = Path.Combine(baseFolder, "downloads");
+            }
+            if (isProd)
+            {
+                await this.GameLocalConfig.SaveConfigAsync(
+                    GameLocalSettingName.ProdDownloadPath,
+                    downloadBaseFolder
+                );
+                await this.GameLocalConfig.SaveConfigAsync(
+                    GameLocalSettingName.ProdDownloadFolderDone,
+                    "False"
+                );
+                await this.GameLocalConfig.SaveConfigAsync(
+                    GameLocalSettingName.ProdDownloadVersion,
+                    previous.Version
+                );
+            }
+            DownloadUpdateFolderConfig folderConfig = new();
+            await GameEventPublisher.PublisAsync(
+                GameContextActionType.CdnSelect,
+                "正在选择最优CDN"
+            );
+            var cdnResult = await TestCdnAsync(
+                _launcher.ResourceDefault.CdnList,
+                previous.BaseUrl,
+                _patch.Resource
+            );
+            if (cdnResult == null)
+            {
+                this.GameEventPublisher.Publish(
+                    new GameContextOutputArgs()
+                    {
+                        Type = GameContextActionType.TipMessage,
+                        TipMessage = "未找到可用的CDN地址，无法进行下载",
+                    }
+                );
+                return false;
+            }
+            var baseUrl = cdnResult.Value.Url + previous.BaseUrl;
+            #endregion
+
+            #region 初始化步骤显示
+            var downloadTasks =
+                new List<(IEnumerable<IndexResource> Items, string Name, string Folder)>();
+            if (patchResource.Any())
+            {
+                this.Setups.Add("下载补丁文件");
+                folderConfig.PatchFolder = Path.Combine(downloadBaseFolder, "patchs");
+                downloadTasks.Add((patchResource, "下载补丁文件", folderConfig.PatchFolder));
+            }
+            if (groupResource.Any())
+            {
+                this.Setups.Add("下载补丁组文件");
+                folderConfig.PatchGroupFolder = Path.Combine(downloadBaseFolder, "patchGroup");
+                downloadTasks.Add((groupResource, "下载补丁组文件", folderConfig.PatchGroupFolder));
+            }
+            if (zipResource.Any())
+            {
+                this.Setups.Add("下载压缩包更新文件");
+                folderConfig.ZipFolder = Path.Combine(downloadBaseFolder, "zips");
+                downloadTasks.Add((zipResource, "下载压缩包更新文件", folderConfig.ZipFolder));
+            }
+            if (downloadResource.Any())
+            {
+                //下载更新文件放最后，直接更新到本体路径上，后续需要测试解压增量是否会影响到downloadResource
+                this.Setups.Add("下载更新文件");
+                folderConfig.DownloadFolder = baseFolder;
+                downloadTasks.Add((downloadResource, "下载更新文件", folderConfig.DownloadFolder));
+            }
+            #endregion
+
+            #region  下载资源
+            for (int i = 0; i < downloadTasks.Count; i++)
+            {
+                var downloadMethod = new DownloadAndVerifyResource(this.Logger)
+                {
+                    ProgressName = downloadTasks[i].Name,
+                };
+
+                downloadMethod.SetParam(
+                    new Dictionary<string, object>()
+                    {
+                        { "resource", downloadTasks[i].Items },
+                        { "launcher", _launcher },
+                        { "isDelete", false },
+                        { "folder", downloadTasks[i].Folder },
+                        { "httpClient", HttpClientService! },
+                        { "downloadState", _downloadState! },
+                        { "baseUrl", baseUrl },
+                        { "isProd", isProd },
+                    },
+                    this.GameEventPublisher
+                );
+                this._currentRunningAction = downloadMethod;
+                CurrentSetups = i;
+                await this.GameEventPublisher.PublishStepAsync(
+                    downloadTasks[i].Name,
+                    CurrentSetups,
+                    Setups
+                );
+                await downloadMethod.ExecuteAsync(true);
+            }
+            #endregion
+
+            #region 安装资源
+            if (isProd) //如果是预下载则跳出
+            {
+                await this.GameLocalConfig.SaveConfigAsync(
+                    GameLocalSettingName.ProdDownloadFolderDone,
+                    "True"
+                );
+                this.GameEventPublisher.Publish(
+                    new GameContextOutputArgs() { Type = GameContextActionType.None }
+                );
+                return true;
+            }
+            else
+            {
+                await this.StartInstallGameResource(_launcher, previous, _patch);
+            }
+            #endregion
+            return true;
+        }
+        catch (TaskCanceledException)
+        {
+            return false;
+        }
+        catch (Exception)
+        {
+            return false;
+        }
+        finally
+        {
+            _actionLock.Release();
+        }
+    }
+
+    /// <summary>
+    /// 开始安装游戏资源
+    /// </summary>
+    /// <param name="launcher"></param>
+    /// <param name="previous"></param>
+    /// <param name="patch"></param>
+    /// <param name="isProd"></param>
+    /// <returns></returns>
+    public async Task StartInstallGameResource(
+        GameLauncherSource launcher,
+        PatchConfig previous,
+        PatchIndexGameResource patch,
+        bool isProd = false
+    )
+    {
+        #region 获取资源
+        var baseFolder = await this.GameLocalConfig.GetConfigAsync(
+            GameLocalSettingName.GameLauncherBassFolder
+        );
+        if (baseFolder == null)
+        {
+            GameEventPublisher.Publish(
+                new GameContextOutputArgs()
+                {
+                    Type = Models.Enums.GameContextActionType.TipMessage,
+                    TipMessage = "未找到游戏安装路径，无法进行安装",
+                }
+            );
+            GameEventPublisher.Publish(
+                new GameContextOutputArgs() { Type = Models.Enums.GameContextActionType.None }
+            );
+
+            return;
+        }
+        #endregion
+
+        DownloadUpdateFolderConfig folderConfig = new();
+        #region 初始化资源
+        this._downloadState = new DownloadState();
+
+        var downloadResource = new List<IndexResource>();
+        var patchResource = new List<IndexResource>();
+        var groupResource = new List<IndexResource>();
+        var zipResource = new List<IndexResource>();
+        foreach (var x in patch.Resource)
+        {
+            if (x.Dest.Contains("krdiff"))
+                patchResource.Add(x);
+            else if (x.Dest.Contains("krpdiff"))
+                groupResource.Add(x);
+            else if (x.Dest.Contains("krzip"))
+                zipResource.Add(x);
+            else
+                downloadResource.Add(x);
+        }
+        string downloadBaseFolder = "";
+        if (isProd)
+        {
+            downloadBaseFolder = Path.Combine(baseFolder, "prodDownloads");
+        }
+        else
+        {
+            downloadBaseFolder = Path.Combine(baseFolder, "downloads");
+        }
+        if (isProd)
+        {
+            await this.GameLocalConfig.SaveConfigAsync(
+                GameLocalSettingName.ProdDownloadPath,
+                downloadBaseFolder
+            );
+            await this.GameLocalConfig.SaveConfigAsync(
+                GameLocalSettingName.ProdDownloadFolderDone,
+                "False"
+            );
+            await this.GameLocalConfig.SaveConfigAsync(
+                GameLocalSettingName.ProdDownloadVersion,
+                previous.Version
+            );
+        }
+        await GameEventPublisher.PublisAsync(GameContextActionType.CdnSelect, "正在选择最优CDN");
+        #endregion
+        Setups.Clear();
+        #region 初始化步骤显示
+        var installTasks =
+            new List<(
+                IEnumerable<IndexResource> Items,
+                string Name,
+                string Folder,
+                InstallGameResourceType
+            )>();
+        if (patchResource.Any())
+        {
+            this.Setups.Add("安装补丁文件");
+            folderConfig.PatchFolder = Path.Combine(downloadBaseFolder, "patchs");
+            installTasks.Add(
+                (
+                    patchResource,
+                    "安装补丁文件",
+                    folderConfig.PatchFolder,
+                    InstallGameResourceType.Krdiff
+                )
+            );
+        }
+        if (groupResource.Any())
+        {
+            this.Setups.Add("安装补丁组文件");
+            folderConfig.PatchGroupFolder = Path.Combine(downloadBaseFolder, "patchGroup");
+            installTasks.Add(
+                (
+                    groupResource,
+                    "安装补丁组文件",
+                    folderConfig.PatchGroupFolder,
+                    InstallGameResourceType.KrdiffGroup
+                )
+            );
+        }
+        if (zipResource.Any())
+        {
+            this.Setups.Add("安装压缩包");
+            folderConfig.ZipFolder = Path.Combine(downloadBaseFolder, "zips");
+            installTasks.Add(
+                (zipResource, "安装压缩包", folderConfig.ZipFolder, InstallGameResourceType.KrZip)
+            );
+        }
+        this.Setups.Add("重新校验文件");
+        folderConfig.DownloadFolder = baseFolder;
+        var resource = await this.GetGameResourceAsync(launcher.ResourceDefault);
+        if (resource != null)
+        {
+            installTasks.Add(
+                (
+                    resource!.Resource,
+                    "安装压缩包",
+                    baseFolder,
+                    InstallGameResourceType.CheckAllFiles
+                )
+            );
+        }
+        else
+        {
+            Logger.WriteError("获取资源信息失败，最终校验启动失败，跳过此校验");
+        }
+
+        for (int i = 0; i < installTasks.Count; i++)
+        {
+            CurrentSetups = i;
+            await this.GameEventPublisher.PublishStepAsync(
+                installTasks[i].Name,
+                CurrentSetups,
+                Setups
+            );
+            await this.GameEventPublisher.PublishStepAsync(
+                installTasks[i].Name,
+                CurrentSetups,
+                Setups
+            );
+            if (installTasks[i].Item4 == InstallGameResourceType.Krdiff)
+            {
+                InstallKrdiffResource installMethod = new InstallKrdiffResource(this.Logger);
+                installMethod.SetParam(
+                    new()
+                    {
+                        { "krdiffs", patchResource },
+                        { "diffFolderPath", installTasks[i].Folder },
+                        { "gameBaseFolder", baseFolder },
+                    },
+                    this.GameEventPublisher
+                );
+                this._currentRunningAction = installMethod;
+                await installMethod.RunAsync();
+            }
+            if (installTasks[i].Item4 == InstallGameResourceType.KrdiffGroup)
+            {
+                InstallKrdiffGroupResource installgroupMethod = new InstallKrdiffGroupResource(
+                    this.Logger
+                );
+                installgroupMethod.SetParam(
+                    new()
+                    {
+                        { "krpdiffs", groupResource },
+                        { "diffFolderPath", installTasks[i].Folder },
+                        { "baseFolderPath", baseFolder },
+                        { "groupFileInfos", patch.GroupInfos },
+                    },
+                    this.GameEventPublisher
+                );
+                this._currentRunningAction = installgroupMethod;
+                //运行解压补丁组后，获取解压结果
+                var newFiles = await installgroupMethod.ExecuteAsync();
+                this.GameEventPublisher.Publish(
+                    new GameContextOutputArgs()
+                    {
+                        Type = GameContextActionType.BottomText,
+                        TipMessage = "正在移动解压文件……",
+                    }
+                );
+                if (newFiles is Dictionary<string, string> newFilesDict)
+                {
+                    var keys = newFilesDict.Keys.ToList();
+                    for (int f = 0; f < keys.Count; f++)
+                    {
+                        string key = keys[i];
+                        string value = newFilesDict[key];
+                        if (File.Exists(value))
+                            File.Delete(value);
+                        File.Move(key, value);
+                    }
+                }
+                else
+                {
+                    await this.StopCannelTaskAsync();
+                    this.GameEventPublisher.Publish(
+                        new GameContextOutputArgs()
+                        {
+                            Type = GameContextActionType.TipMessage,
+                            TipMessage = "补丁组安装失败，无法进行后续安装，请尝试直接修复文件",
+                        }
+                    );
+                    return;
+                }
+            }
+            if (installTasks[i].Item4 == InstallGameResourceType.KrZip) { }
+            if (installTasks[i].Item4 == InstallGameResourceType.CheckAllFiles)
+            {
+                var downloadMethod = new DownloadAndVerifyResource(this.Logger)
+                {
+                    ProgressName = "重新校验文件",
+                };
+                await GameEventPublisher.PublisAsync(
+                    GameContextActionType.CdnSelect,
+                    "正在选择最优CDN"
+                );
+                var cdnResult = await TestCdnAsync(
+                    launcher.ResourceDefault.CdnList,
+                    previous.BaseUrl,
+                    patch.Resource
+                );
+                if (cdnResult == null)
+                {
+                    Logger.WriteError("获取资源信息失败，最终校验启动失败，跳过此校验");
+                    this.GameEventPublisher.Publish(
+                        new GameContextOutputArgs()
+                        {
+                            Type = GameContextActionType.None,
+                        }
+                    );
+                    return;
+                }
+                var baseUrl = cdnResult.Value.Url + previous.BaseUrl;
+                downloadMethod.SetParam(
+                    new Dictionary<string, object>()
+                    {
+                        { "resource", installTasks[i].Items.ToList() },
+                        { "launcher", launcher },
+                        { "isDelete", false },
+                        { "folder", installTasks[i].Folder },
+                        { "httpClient", HttpClientService! },
+                        { "downloadState", _downloadState! },
+                        { "baseUrl", baseUrl },
+                        { "isProd", isProd },
+                    },
+                    this.GameEventPublisher
+                );
+                this._currentRunningAction = downloadMethod;
+                CurrentSetups = i;
+            }
+        }
+        var writeConfig = new WriteGameResourceConfig(
+               this.GameLocalConfig,
+               launcher,
+               this.Config,
+               Logger
+           );
+        await writeConfig.WriteDownloadAndUpDateResultAsync(launcher);
+        #endregion
+    }
+    #endregion
+}
